requirements.txt
Flask==2.1.0
pymongo==4.0.1
python-dotenv==0.19.1
